+++ src/Dependencies/Diagnostics/DiagnosticSeverity.cs
@@ -5,7 +5,7 @@ namespace Diagnostics;
 /// Severity of <see cref="Diagnostic" />,
 /// does not effect how the <see cref="DiagnosticQueue<T>" /> interacts with them.
 /// </summary>
-public enum DiagnosticSeverity : int {
+public enum DiagnosticSeverity : byte {
     All = 0,
     Debug = 1,
     Info = 2,
